LayoutTests/http/tests/dom/iframe-no-src-set-location-expected.txt

This tests that an iframe with no src attribute uses the parent document's URL as a base when resolving non-fully-qualified URLs set by JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('theFrame').contentDocument.baseURI is "http://127.0.0.1:8000/dom/iframe-no-src-set-location.html"
PASS document.getElementById('theFrame').contentDocument.body.innerHTML is 'PASS\n'
PASS successfullyParsed is true

TEST COMPLETE
